+++ dg/commands/fyre/cluster/create_for_db2_data_gate.py
@@ -24,7 +24,7 @@ import dg.config
 import dg.lib.click.utils
 import dg.utils.network
 
-from dg.lib.fyre.api_manager import OCPPlusAPIManager
+from dg.lib.fyre.ocp_plus_api_manager import OCPPlusAPIManager
 from dg.lib.fyre.data.ocpplus_cluster_specification import (
     HAProxyTimoutSettings,
     OCPPlusClusterSpecification,
