+++ dg/commands/fyre/cluster/disable_delete.py
@@ -21,7 +21,7 @@ import dg.config.cluster_credentials_manager
 import dg.lib.click.utils
 import dg.utils.network
 
-from dg.lib.fyre.api_manager import OCPPlusAPIManager
+from dg.lib.fyre.ocp_plus_api_manager import OCPPlusAPIManager
 from dg.lib.fyre.utils.click import fyre_command_options
 from dg.utils.logging import loglevel_command
