+++ dg/commands/fyre/login.py
@@ -16,7 +16,7 @@ import dg.config
 import dg.utils.network
 
 from dg.lib.error import DataGateCLIException
-from dg.lib.fyre.api_manager import OCPPlusAPIManager
+from dg.lib.fyre.ocp_plus_api_manager import OCPPlusAPIManager
 from dg.lib.fyre.utils.click import fyre_command_options
 from dg.utils.logging import loglevel_command
