+++ dg/utils/http_method.py
@@ -1,0 +1,14 @@
+from enum import Enum, auto
+
+
+class HTTPMethod(Enum):
+    DELETE = auto()
+    GET = auto()
+    HEAD = auto()
+    OPTIONS = auto()
+    PATCH = auto()
+    POST = auto()
+    PUT = auto()
+
+    def __str__(self):
+        return self.name
