+++ dg/utils/network.py
@@ -18,7 +18,7 @@ import socket
 from typing import List
 
 import netifaces
-import requests
+import urllib3
 
 
 def disable_insecure_request_warning():
@@ -28,7 +28,7 @@ def disable_insecure_request_warning():
     # InsecureRequestWarning: Unverified HTTPS request is being made to host
     # '…'. Adding certificate verification is strongly advised. See:
     # https://urllib3.readthedocs.io/en/latest/advanced-usage.html#ssl-warnings
-    requests.packages.urllib3.disable_warnings(category=requests.packages.urllib3.exceptions.InsecureRequestWarning)
+    urllib3.disable_warnings(category=urllib3.exceptions.InsecureRequestWarning)
 
 
 def get_local_ipv4_addresses() -> List[ipaddress.IPv4Address]:
